feat: update delete create sub product optional

+++ src/api/lib/models/Store/cat.js
@@ -1,4 +1,4 @@
-import { INTEGER, STRING, TEXT, TINYINT, literal } from 'sequelize'
+import { INTEGER, STRING, TEXT, literal } from 'sequelize'
 import connect from '../../db'
 const sequelize = connect()
 import { enCode } from '../../utils/util'
@@ -46,7 +46,7 @@ const catProducts = sequelize.define('catproducts', {
     allowNull: true
   },
   pState: {
-    type: TINYINT,
+    type: INTEGER,
     allowNull: true
   },
   pDatCre: {
