+++ src/api/lib/models/product/productFood.js
@@ -1,4 +1,4 @@
-import { INTEGER, STRING, TEXT, TINYINT, literal } from 'sequelize'
+import { INTEGER, STRING, TEXT, literal } from 'sequelize'
 import connect from '../../db'
 const sequelize = connect()
 import SizeModel from '../information/size'
@@ -175,11 +175,11 @@ const productModelFood = sequelize.define('productmodelfood', {
     allowNull: true
   },
   pState: {
-    type: TINYINT,
+    type: INTEGER,
     allowNull: false
   },
   sTateLogistic: {
-    type: TINYINT,
+    type: INTEGER,
     allowNull: false
   },
   // Si el producto esta asegurado ( Protegido )
